Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -16,7 +16,7 @@ int main(){
     }
     else if(m== '/'){
         if(b!=0){
-            y=(float)a/b;
+            y=a/b;
         }
         else{
             printf("Error\n");
